requirements.txt
appdirs==1.4.4
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cmake==3.30.2
contourpy==1.3.0
cycler==0.12.1
docker-pycreds==0.4.0
einops==0.7.0
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
ftfy==6.2.3
gitdb==4.0.11
GitPython==3.1.43
h5py==3.9.0
huggingface-hub==0.17.3
idna==3.10
imageio==2.35.1
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
lazy_loader==0.4
lightning-utilities==0.11.6
lit==18.1.8
MarkupSafe==2.1.5
matplotlib==3.7.3
mpmath==1.3.0
networkx==3.3
numpy==1.25.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
nystrom-attention==0.0.12
open-clip-torch==2.23.0
opencv-python==4.8.0.76
openslide-python==1.3.1
packaging==24.1
pandas==2.0.3
pathtools==0.1.2
pillow==10.4.0
pip==24.2
protobuf==4.25.5
psutil==6.0.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
safetensors==0.4.4
scikit-image==0.22.0
scikit-learn==1.3.0
scipy==1.14.1
sentencepiece==0.2.0
sentry-sdk==2.13.0
setproctitle==1.3.3
setuptools==75.1.0
six==1.16.0
smmap==5.0.1
sympy==1.13.2
threadpoolctl==3.5.0
tifffile==2024.9.20
timm==0.9.2
tokenizers==0.14.1
torch==2.0.0
torchaudio==2.0.1
torchmetrics==1.1.0
torchvision==0.15.1
tqdm==4.66.5
transformers==4.34.1
triton==2.0.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wandb==0.15.8
wcwidth==0.2.13
wheel==0.44.0
